send product sistemi tamamlandı.

+++ editor.tshirthane.com/app/resources/views/admin/type/urun-sablonlari.blade.php
@@ -1,12 +1,70 @@
 <div class="content">
     <div class="row">
-        <?php 
-        $table_name = str_slug($c->title,'_');
-        if(getisset("edit"))  { 
-            $edit = db($table_name)->where("id",get("edit"))->first();
-          ?>
-                @include("admin.type.urun-sablonlari.edit-form")
+            <div class="col-12">
+                <?php 
+                    $table_name = str_slug($c->title,'_');
+                    if(getisset("edit"))  { 
+
+                        if(getisset("update")) {
+                            $p = $_POST;
+                        //  dump($p);
+                            $json = $p;
+                            $file = post("old_file");
+                            $json['longDescription'] = $json['html'];
+                            $json['shortDescription'] = substr($json['html'],0,150);
+                            unset($json['html']);
+                            unset($json['_token']);
+                            unset($json['old_file']);
+                            unset($json['oldslug']);
+                            unset($json['id']);
+                            unset($json['kid']);
+                            unset($json['type']);
+                            
+                            //dump($json);
+                            if($_FILES['file']['name']!="") {
+                                $file = upload("file","mockup/");
+                            //   echo $file;
+                            }
+                            $file = str_replace("//","/",$file);
+                            $json['images'][0]['imageUrl'] = env('APP_URL') . $file;
+                            dump($json);
+                            db($table_name)->where("id",get("edit"))->update([
+                                "title" => $p['title'],
+                                "html" => $p['html'],
+                                "files" => $file,
+                                "json" => json_encode_tr($json)
+                            ]);
+                        //  dump($json);
+                            bilgi("Güncelleme yapıldı");
+                            //exit();
+                        }
+                        
+                        
+
+                        $edit = db($table_name)->where("id",get("edit"))->first();
+
+                        if($edit->json!="") {
+                            $j = j($edit->json);
+                        //   dump($j);
+
+                            if(getisset("test")) {
+                                $data = $j;
+                                //dump($data);
+                                sendProduct($data);
+                                 ?>
+                                 <a href="https://www.tshirthane.com/template/common/api_docs/web_servis_dokuman_v2.0.4.pdf" class="btn btn-success">API Dokümantasyonu</a>
+
+                                 <?php 
+                                exit();
+                            }
+                        }
+
+                    ?>
+            </div>
+            
+            @include("admin.type.urun-sablonlari.edit-form")
          <?php } ?>
+    </div>
     </div>
 
 <div class="block">
@@ -24,44 +82,36 @@
         <div class="block-content">
 			<div class="js-gallery ">
 			@include("admin.inc.table-search")
-			<div class="table-responsive">
-            <table class="table table-striped table-hover table-bordered table-vcenter">
-                <thead>
-                    <tr>
-                        <th>{{__("Template Name")}}</th>
-						
-                        <th class="text-center" style="width: 100px;">{{__("Transaction")}}</th>
-                    </tr>
-                </thead>
-                <tbody>
-				<?php $alt = db($table_name)->orderBy("id","DESC")->simplePaginate(20); ?>
+            <?php 
+            if(getisset("delete")) {
+                    db($table_name)->where("id",get("delete"))->delete();
+                    bilgi(get("id") . " ürün şablonu silinmiştir.");
+                }  
+            ?>
+            <div class="row">
+                <?php 
+                  
+                $alt = db($table_name)->orderBy("id","DESC")->simplePaginate(20); ?>
 				@foreach($alt AS $a)
-					<tr class="">
-                       
-                        <td>
-                            {{$a->title}}
-                            
-						</td>
-						
-						
-					  <td class="text-center">
-                            <div class="btn-group">
-                                <a href="?edit={{$a->id}}" class="btn btn-secondary js-tooltip-enabled" data-toggle="tooltip" title="" data-original-title="Edit">
-                                    <i class="fa fa-edit"></i>
-                                </a>
-                                <a href="?delete={{$a->id}}" teyit="{{$a->title}} {{__('içeriğini silmek istediğinizden emin misiniz?')}}" title="{{$a->title}} {{__('Silinecek!')}}" class=" btn  btn-secondary js-tooltip-enabled" data-toggle="tooltip" title="" data-original-title="Delete">
-                                    <i class="fa fa-times"></i>
-                                </a>
-                            </div>
-                        </td>
-                    </tr>
-				@endforeach
-				
-                                     
-                                    </tbody>
-            </table>
+                <?php if($a->title=="") $a->title="Adsız şablon"; ?>
+                 {{col("col-md-3",$a->title)}} 
+                    <div style="width:250px;height:250px;">
+                        <img src="{{picture2($a->files,256,0)}}" style=""  alt="">
+                    </div>
+                    <div class="btn-group">
+                        <a href="?edit={{$a->id}}" class="btn btn-secondary js-tooltip-enabled" data-toggle="tooltip" title="" data-original-title="Edit">
+                            <i class="fa fa-edit"></i>
+                        </a>
+                        <a href="?delete={{$a->id}}" teyit="{{$a->title}} {{__('içeriğini silmek istediğinizden emin misiniz?')}}" title="{{$a->title}} {{__('Silinecek!')}}" class=" btn  btn-secondary js-tooltip-enabled" data-toggle="tooltip" title="" data-original-title="Delete">
+                            <i class="fa fa-times"></i>
+                        </a>
+                    </div>
+                 {{_col()}}
+                 @endforeach
+            </div>
+			
 			{{$alt->fragment('alt')->links() }}
-			</div>
+		
 			</div>
         </div>
